refactor: streamline Action type definition and update Block component to use new Action type
