+++ frontend/src/types/Round.ts
@@ -1,27 +1,4 @@
-import { IFeedbackForm } from "@/components/Trial/Trial";
-import { PlaybackArgs } from "./Playback";
-import { TrialConfig } from "./Trial";
-
-export interface Action {
-  view: string;
-
-  // Optional properties depending on the view
-  instruction?: string;
-  button_label?: string;
-  steps?: Array<{
-    number: number;
-    description: string;
-  }>;
-  timer?: number | null;
-  title?: string;
-  config?: TrialConfig;
-  style?: {
-    root: string;
-  };
-  playback?: PlaybackArgs;
-  html?: { body: string | TrustedHTML };
-  feedback_form?: IFeedbackForm;
-}
+import { Action } from "./Action";
 
 export type Round = Action[];
